modify care team confirmation context for alerting
Care team invitation/confirmation contexts are currently a
clients.Permissions object. To support alerting, the context is extended to
include a alerts.Config (defined in platform).

The custom JSON unmarshaler allows flexibility in parsing the following
received contexts:

    1. The existing permissions only context
    2. A hybrid context with old-style permissions plus an alertsConfig
    3. A future context with a "Permissions" and an alertsConfig as sibling
       properties

When the API is migrated to scenario #3 above, the custom marshaler can be
removed.

Bringing in platform as a dependency brought many vendor changes as well.

Part of BACK-2500