fix accepting invitations without an alerts config
When adding support for alerts config, a change was made to AcceptInvite to
upsert the alerts config. Which is fine, unless the alerts config is
nil. Brian found this problem during his testing (thanks!)

There were no tests of accepting a valid invite before this. So one has been
added, that accepts an invite without an alerts config, and ensures that the
nil alerts config isn't upserted.

BACK-2500