move to zap logging
During code review, it was requested that the zap logger be used.

I might have gone a bit overboard, but I've moved everything over to zap. I
think it's still TBD if zap is what we really want, but at least now we
can see it in action, and better evaluate if it brings advantages over other
options.

Zap produces a stack trace at the Error log level, so I used that as a
guide. In general, any time a function returned an error, and that error
caused the HTTP handler to abort its processing with a 5XX HTTP status, I used
the Error level for the log, so we'll log the stack trace that led up to the
error.

In the event that an error was returned and the handler returned a 4XX HTTP
status, I used the Info log level. My thinking was if the handler was telling
the client that the error was on their side, then whatever the Go error was,
we didn't need a stack trace. This happens a lot for example when a request
body contains invalid JSON. The Unmarshal call will fail with an error and the
handler returns a HTTP 400 or similar. A stack trace in that event wouldn't be
useful.

In general, I didn't use the Warn level much, as it behaves differently in
prod versus non-prod. In the former, it doesn't include a stack trace, while
in the latter it does. I used it during an optional permissions check. In
development, we'll get better debug info, but in production we'll skip the
stack trace, as the error isn't going to cause an abort of the handler method.

I also refactored many API responses to use the existing sendError and other
related methods. This should help maintain consistency in the future, and
should we decide to remove zap for example, we'll have fewer places to modify,
as the move to using sendError and friends has greatly reduced the number of
log calls that were scattered around the code base. (Last I checked, the
number of logging sites was reduced from 153 to 49, according to rough grep
checks.)

Another request from code review, was that logging should include the userId
where one existed. This was accomplished using a middleware function, such
that the userId is added to the logger's context and will automatically appear
in the log output. There's no need to manually specify it at the logging call
site.

The zap dependency version was updated to 1.22 to get access to the
WithOptions method of the Sugared logger, which makes some of our logging
cleaner.

BACK-2500