don't ignore errors, or leak context timeouts
A number of errors were being ignored. Some were pretty inocuous, but all are
easy to fix.

These changes will make it easier to migrate to the zap logger. With better
error control, the number of places in the code that need to log versus
simply returning an error is reduced. As the number of places that need to
log is reduced, it's easier to thread loggers into the stack.

Some of the modifications required changing functions to return errors where
they previously were being ignored. In some of those places logging was lifted
up into the calling function, leaving the inner function without a dependency
on logging.

Of the ignored errors, some would have led to nil pointer exceptions, so these
are worth fixing, as now the service won't crash. Instead, we'll have a logged
error message that we can use to help diagnose the issue.

BACK-2500