remove findExistingConfirmation
The function was writing HTTP responses and logging, both of which are better
done in the actual HTTP handler methods. When these helper methods know about
HTTP and can (but don't always) write responses, it makes the handlers awkward
to write, because it's never clear whether they should just return afterwards,
or if they can do any error recovery or better logging, etc.

Without the logging and sometimes writing of HTTP responses, the helper wasn't
doing anything, and so it's easily removed.

Also, the error handling of FindConfirmation is made consistent as an internal
server error.

BACK-2500