rework findResetConfirmation
The function was writing HTTP responses and logging, both of which are better
done in the actual HTTP handler methods. When these helper methods know about
HTTP and can (but don't always) write responses, it makes the handlers awkward
to write, because it's never clear whether they should just return afterwards,
or if they can do any error recovery or better logging, etc.

As one example, findResetConfirmation would write 404 responses if the
confirmation wasn't found. The handler ignored that one particular case, and
it wasn't obvious why it was ignoring it. Even as the handler ignored it, it
still had to have a nil check/guard statement, which indented the code
awkwardly and left the reader wondering just what was returned the user if
conf was nil.

BACK-2500