fix Api logging thread-safety issue
When migrating to zap logging, the Api.logger was modified by middleware that
was adding the request's userID to the logger's context in a way that wasn't
thread-safe. The result was that many many non-sensical userIDs were logged in
the context, producing log output that wasn't at all useful.

The best way to fix this was to properly generate a child logger for each
request with the appropriate userID in its logging context, then pass that
logger into the handlers via their request's context.

As a result, the fix touches a lot of code—anywhere that logs, now needs to
retrieve the request-specific logger from that request's context.

Tests were written that attempt to exacerbate the problem, but as its a caused
by race conditions, it's very hard to reliably reproduce. Flappiness in the
test is a red flag though, and comments were added to indicate that.

BACK-2500