add clinicId to the logging context
As requested in code review. The addUserIDToLogger method was made
general-purpose. In addition, instead of using it only on those handlers that
have a userId/clinicId parameter, it's used as a generic middleware on all
requests. This small performance hit makes it so that any future endpoints
that are added will automatically log userid and clinicid, plus the endpoint
definition code is a bit cleaner, as it's not having to call uid helpers all
over the place.

BACK-2500