ignore a vulnerabilty found by snyk
This vulnerability is a buffer overflow in a protocol buffer library used by a
dependency. Our services don't use protocol buffers, so we shouldn't be using
whatever part of that has the vulnerability.

The full path to the vulnerability is:

github.com/tidepool-org/clinic/client@#ad2e71e79f6f
> github.com/deepmap/oapi-codegen/pkg/runtime@1.16.2
> github.com/kataras/iris/v12@12.2.7
> github.com/kataras/iris/v12/x/errors@12.2.7
> github.com/kataras/iris/v12/macro/handler@12.2.7
> github.com/kataras/iris/v12/context@12.2.7
> google.golang.org/protobuf/encoding/protojson@1.31.0